requirements.txt
grpcio~=1.66.1
protobuf~=5.28.2
simcash~=0.1.0
setuptools~=75.1.0